Initial module (incomplete) for logging config
Adds the `phlawg.config` module with a `from_environment` function
for configuring the `logging` package based on environment variables
and phlawg-friendly defaults.

Initial commit just gives the defaults plus an ability to specify
the packages you expect to use for metrics.  Uses the json logger
for metrics, basic format for regular logs, everything going to
STDERR.